separate as-path-list
