+++ catalystwan/models/policy/__init__.py
@@ -8,6 +8,7 @@ from typing_extensions import Annotated
 
 from catalystwan.models.policy.list.app import AppList
 from catalystwan.models.policy.list.app_probe import AppProbeClassList
+from catalystwan.models.policy.list.as_path import ASPathList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -25,7 +26,6 @@ from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
 from .lists import (
-    ASPathList,
     ClassMapList,
     ColorList,
     CommunityList,
