+++ catalystwan/models/policy/list/as_path.py
@@ -1,0 +1,18 @@
+from typing import List, Literal
+
+from catalystwan.models.policy.lists import PolicyListBase
+from catalystwan.models.policy.lists_entries import ASPathListEntry
+from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+
+
+class ASPathList(PolicyListBase):
+    type: Literal["asPath"] = "asPath"
+    entries: List[ASPathListEntry] = []
+
+
+class ASPathListEditPayload(ASPathList, PolicyListId):
+    pass
+
+
+class ASPathListInfo(ASPathList, PolicyListInfo):
+    pass
